third_party/WebKit/LayoutTests/external/wpt/xhr/setrequestheader-allow-whitespace-in-value-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: setRequestHeader() - header value with whitespace ( ) assert_equals: expected "X-Empty: \n" but got "Syntax error: no colon and space found: X-Empty:"
PASS XMLHttpRequest: setRequestHeader() - header value with whitespace ( t)
PASS XMLHttpRequest: setRequestHeader() - header value with whitespace (t )
PASS XMLHttpRequest: setRequestHeader() - header value with whitespace ( t )
Harness: the test ran to completion.
